LINQ-19: Add change-tracking to BucketContext
Motivation
----------
This commit introduces experimental support for change tracking of
entities/POCOs loaded from documents stored in Couchbase server. By enabling
change tracking through a call to the BeginChangeTracking method on
BucketContext, each POCO or entity that corresponds to a document returned
by a query will be intercepted and proxied. The proxy contains logic for
detecting any modifications to a document and adding the corresponding
object to a mutated list. When SubmitChanges is called, the changed
documents will be added, updated or deleted on Couchbase Server. Change
tracking be disabled by calling the EndChangeTracking method; which will
remove any modified documents from the mutated list and subsequent queries
will not be intercepted and proxied.

Modifications
-------------
The BucketContext is now a subcriber for all proxy modifications when
change tracking is enabled; any changes to an object which is proxied will
be bubbled up to the BucketContext and stored in a mutated list. If a new
document is added that was not proxied it will be wrapped in a
NewDocumentWrapper class and treated as if it was proxied. Various other
methods of BucketContext were also modified to support the cases where
chnage tracking is enabled and when it is not enabled. In general, any
delete, add or mutation persistence will be deferred until SubmitChanges
is called.

Results
-------
When BegingChangeTracking is enabled all documents returned by a query
will be proxied. If a document is modified, the BucketContext will be
notified. When SubmitChanges it called, the documents that have been
changed or added will be persisted to Couchbase.